Add Google search and MSN search

+++ Tranzact.SearchFight.API.Entities/OUTPUT/SearchOUT.cs
@@ -8,6 +8,6 @@ namespace Tranzact.SearchFight.API.Entities.OUTPUT
     {
         public string engine { get; set; }
         public string word { get; set; }
-        public int totalRecords { get; set; }
+        public long totalRecords { get; set; }
     }
 }
